html/rendering/widgets/baseline-alignment-and-overflow.tentative.html

<!DOCTYPE html>
<title>HTML: widgets' baseline alignment and interaction with 'overflow'</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
.test {
  border-collapse: collapse;
  font-size: 10px;
}
.test td {
  border: none;
  padding: 0;
  margin:0;
  outline: 1px solid silver;
}

.test td > input,
.ref td > .fake-input-text,
.ref td > .inline-block,
.ref td > img,
.ref td > button {
  font: inherit;
  height: 60px;
  width: 60px;
  padding: 10px;
  box-sizing: border-box;
  margin: 10px 0;
}
.ref button img {
  height: 100%;
  width: 100%;
  display: block;
}
.ref .fake-input-text {
  display: inline-grid;
  border: 2px solid;
  align-items: center;
}
.ref .inline-block {
  display: inline-block;
}
.ref td > img.no-margin-bottom {
  margin-bottom: 0;
}
[style*="appearance: none;"] {
  -webkit-appearance: none; /* TODO(zcorpan) remove this when unprefixed appearance is supported */
}
</style>
<div id="log"></div>
<h2>refs</h2>
<!--
The first span's offsetTop is what we want to compare with the corresponding test's span's offsetTop.
The sibling element is there to control where the baseline for the line box will be.
-->
<table class="test ref">
  <tr class="ref-text"><td><span>ref for text</span> <span class=fake-input-text>x</i>
  <tr class="ref-checkbox"><td><span>ref for checkbox</span> <img class=no-margin-bottom src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABgAAAAMCAIAAAD3UuoiAAAAGklEQVQoz2Nk%2BP%2BfgRqAiYFKYNSgUYOGp0EA%2BQMCFrJdTgsAAAAASUVORK5CYII%3D">
  <tr class="ref-color"><td><span>ref for color</span> <button><img src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABgAAAAMCAIAAAD3UuoiAAAAGklEQVQoz2Nk%2BP%2BfgRqAiYFKYNSgUYOGp0EA%2BQMCFrJdTgsAAAAASUVORK5CYII%3D"></button>
  <tr class="ref-file"><td><span>ref for file</span> <span class=inline-block><button>x</button></span>
  <tr class="ref-image-alt"><td><span>ref for image alt</span> <span class=inline-block>x</span>
  <tr class="ref-image-src"><td><span>ref for image src</span> <img src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABgAAAAMCAIAAAD3UuoiAAAAGklEQVQoz2Nk%2BP%2BfgRqAiYFKYNSgUYOGp0EA%2BQMCFrJdTgsAAAAASUVORK5CYII%3D">
</table>
<h2>template table</h2>
<!--
Each row in this table will be cloned into #test-table for each combination of styles to test.
-->
<table id="template-table">
  <tr><td><span>text</span> <input type=text value=x></td></tr>
  <tr><td><span>search</span> <input type=search value=x></td></tr>
  <tr><td><span>tel</span> <input type=tel value=x></td></tr>
  <tr><td><span>url</span> <input type=url value="data:,x"></td></tr>
  <tr><td><span>email</span> <input type=email value=x></td></tr>
  <tr><td><span>password</span> <input type=password value=x></td></tr>
  <tr><td><span>date</span> <input type=date value="2020-01-01"></td></tr>
  <tr><td><span>month</span> <input type=month value="2020-01"></td></tr>
  <tr><td><span>week</span> <input type=week value="2020-W01"></td></tr>
  <tr><td><span>time</span> <input type=time value="00:00"></td></tr>
  <tr><td><span>datetime-local</span> <input type=datetime-local value="2020-01-01T00:00"></td></tr>
  <tr><td><span>number</span> <input type=number value=0></td></tr>
  <tr><td><span>range</span> <input type=range></td></tr>
  <tr><td><span>color</span> <input type=color value=#000000></td></tr>
  <tr><td><span>checkbox</span> <input type=checkbox></td></tr>
  <tr><td><span>radio</span> <input type=radio></td></tr>
  <tr><td><span>file</span> <input type=file></td></tr>
  <tr><td><span>submit</span> <input type=submit value=x></td></tr>
  <tr><td><span>image</span> <input type=image src="data:,broken" alt="x"></td></tr>
  <tr><td><span>image-with-src</span> <input type=image src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABgAAAAMCAIAAAD3UuoiAAAAGklEQVQoz2Nk%2BP%2BfgRqAiYFKYNSgUYOGp0EA%2BQMCFrJdTgsAAAAASUVORK5CYII%3D" alt="x"></td></tr>
  <tr><td><span>reset</span> <input type=reset value=x></td></tr>
  <tr><td><span>button</span> <input type=button value=x></td></tr>
</table>
<h2>tests</h2>
<!--
This table gets populated by the script.
-->
<table class="test" id="test-table"><tbody></tbody></table>
<script>
  "use strict";

  promise_setup(async () => {
    const templateTable = document.querySelector('#template-table');
    const testTBody = document.querySelector('#test-table tbody');

    const templateRows = templateTable.querySelectorAll('tr');
    for (let templateRow of templateRows) {
      for (var appearanceValue of ["auto", "none"]) {
        for (var overflowValue of ['visible', 'hidden', 'scroll']) {
          const clonedRow = templateRow.cloneNode(true);
          clonedRow.querySelector('input').setAttribute('style', `overflow: ${overflowValue}; appearance: ${appearanceValue};`);
          testTBody.append(clonedRow);
        }
      }
    }

    // wait for images to load
    await new Promise(resolve => window.onload = e => resolve());
    for (const img of document.images) {
      assert_precondition(img.complete); // either error state or loaded
    }

    // get layout info from refs
    const refTextOffsetTop = document.querySelector('.ref-text span').offsetTop;
    const refCheckboxOffsetTop = document.querySelector('.ref-checkbox span').offsetTop;
    const refColorOffsetTop = document.querySelector('.ref-color span').offsetTop;
    const refFileOffsetTop = document.querySelector('.ref-file span').offsetTop;
    const refImageAltOffsetTop = document.querySelector('.ref-image-alt span').offsetTop;
    const refImageSrcOffsetTop = document.querySelector('.ref-image-src span').offsetTop;

    function expectedOffsetTop(input) {
      // TODO(zcorpan) https://github.com/whatwg/html/issues/5065
      // for now this is intended to match Firefox
      const style = input.getAttribute('style');
      const src = input.getAttribute('src');
      switch (input.type) {
        case 'file':
          return refFileOffsetTop;
        case 'range':
          return refImageSrcOffsetTop;
        case 'color':
          return refColorOffsetTop;
        case 'checkbox':
        case 'radio':
          return (style.includes('appearance: none;')) ? refImageSrcOffsetTop : refCheckboxOffsetTop;
        case 'file':
          return refFileOffsetTop;
        case 'image':
          return (src === 'data:,broken' && style.includes('overflow: visible;')) ? refImageAltOffsetTop : refImageSrcOffsetTop;
      }
      return refTextOffsetTop;
    }

    function testName(markup) {
      return markup.replace(/data:image\/png[^"]+/, 'data:(png)');
    }

    for (const row of testTBody.children) {
      const input = row.firstChild.lastElementChild;
      const allowedDelta = 3;
      promise_test(async () => {
        assert_precondition(input.type === input.getAttribute('type'), 'input type should be supported')
        const offsetTopActual = row.firstChild.firstChild.offsetTop;
        assert_approx_equals(offsetTopActual, expectedOffsetTop(input), allowedDelta, '<span>.offsetTop');
      }, testName(input.outerHTML));
    }
  });
</script>
