third_party/blink/web_tests/external/wpt/html/rendering/widgets/baseline-alignment-and-overflow.tentative-expected.txt

This is a testharness.js-based test.
Found 16 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] <input type="range" style="overflow: visible; appearance: auto;">
  assert_equals: <span>.offsetTop expected 70 but got 60
[FAIL] <input type="range" style="overflow: hidden; appearance: auto;">
  assert_equals: <span>.offsetTop expected 70 but got 60
[FAIL] <input type="range" style="overflow: scroll; appearance: auto;">
  assert_equals: <span>.offsetTop expected 70 but got 60
[FAIL] <input type="range" style="overflow: visible; appearance: none;">
  assert_equals: <span>.offsetTop expected 70 but got 60
[FAIL] <input type="range" style="overflow: hidden; appearance: none;">
  assert_equals: <span>.offsetTop expected 70 but got 60
[FAIL] <input type="range" style="overflow: scroll; appearance: none;">
  assert_equals: <span>.offsetTop expected 70 but got 60
[FAIL] <input type="color" value="#000000" style="overflow: visible; appearance: auto;">
  assert_equals: <span>.offsetTop expected 48 but got 45
[FAIL] <input type="color" value="#000000" style="overflow: hidden; appearance: auto;">
  assert_equals: <span>.offsetTop expected 48 but got 45
[FAIL] <input type="color" value="#000000" style="overflow: scroll; appearance: auto;">
  assert_equals: <span>.offsetTop expected 48 but got 45
[FAIL] <input type="color" value="#000000" style="overflow: visible; appearance: none;">
  assert_equals: <span>.offsetTop expected 48 but got 45
[FAIL] <input type="color" value="#000000" style="overflow: hidden; appearance: none;">
  assert_equals: <span>.offsetTop expected 48 but got 45
[FAIL] <input type="color" value="#000000" style="overflow: scroll; appearance: none;">
  assert_equals: <span>.offsetTop expected 48 but got 45
[FAIL] <input type="image" src="data:,broken" alt="x" style="overflow: hidden; appearance: auto;">
  assert_equals: <span>.offsetTop expected 20 but got 70
[FAIL] <input type="image" src="data:,broken" alt="x" style="overflow: scroll; appearance: auto;">
  assert_equals: <span>.offsetTop expected 20 but got 70
[FAIL] <input type="image" src="data:,broken" alt="x" style="overflow: hidden; appearance: none;">
  assert_equals: <span>.offsetTop expected 20 but got 70
[FAIL] <input type="image" src="data:,broken" alt="x" style="overflow: scroll; appearance: none;">
  assert_equals: <span>.offsetTop expected 20 but got 70
Harness: the test ran to completion.
